erms-468 cost items

+++ app/grails-app/domain/com/k_int/kbplus/CostItem.groovy
@@ -28,7 +28,7 @@ class CostItem extends BaseDomainComponent {
 
     Double costInBillingCurrencyAfterTax
     Double costInLocalCurrencyAfterTax
-    Double taxRate
+    Integer taxRate
     Boolean finalCostRounding
 
     String costTitle
